# Unify CPU hist sketching