@SmirnovEgorRu @ShvetsKS


This PR merges sparse cut builder and dense cut builder into a single sketch container.  At the same time it removes duplicated set of tests and adds tests for distributed environment.  I haven't optimize sparse dataset on distributed environment yet, so it will still generate OOM error as before.  But for single node, as after this PR we don't need to generate an extra copy of dataset from `GetBatches<CSCPage>()`, memory usage should be quite good for sketching.

- Drive by optimization:  I avoided the call to `operator[]` of `SparsePage` by introducing a data view.  So we won't be needing to call `rabit::IsDistributed()` and `HostVector` at every row access.  If it's a burden for review I can revert it.

## Motivation

I'm trying to add categorical data support in XGBoost.  The first hurdle is sketching.  Currently XGBoost has 7 entry points for sketching, I'm trying to merge as many of them as possible.

## Performance for sketching

* Single node 1 x 16

** Before
*** URL
GmatInitialization: 9.75871s, 1 calls @ 9758709us
GmatInitialization: 9.71126s, 1 calls @ 9711261us
GmatInitialization: 9.80294s, 1 calls @ 9802938us

*** HIGGS
GmatInitialization: 3.74729s, 1 calls @ 3747286us
GmatInitialization: 3.90515s, 1 calls @ 3905151us
GmatInitialization: 3.79345s, 1 calls @ 3793448us

** After
*** URL
GmatInitialization: 10.3033s, 1 calls @ 10303323us
GmatInitialization: 10.2406s, 1 calls @ 10240576us
GmatInitialization: 10.224s, 1 calls @ 10224020us

*** HIGGS
GmatInitialization: 3.74322s, 1 calls @ 3743224us
GmatInitialization: 3.63079s, 1 calls @ 3630793us
GmatInitialization: 3.67905s, 1 calls @ 3679049us

* 4 x 4

*** HIGGS
** Before
GmatInitialization: 3.83843s, 1 calls @ 3838427us
GmatInitialization: 3.89644s, 1 calls @ 3896437us
GmatInitialization: 3.90035s, 1 calls @ 3900354us

** After
GmatInitialization: 3.7682s, 1 calls @ 3768198us
GmatInitialization: 3.707s, 1 calls @ 3707000us
GmatInitialization: 3.70828s, 1 calls @ 3708276us

URL for distributed is not tested as it will blow up the memory.  I have another branch for optimizing it, the target is we can unify the sketching between `hist` and `approx`.