@trivialfis 2020-07-15
Interestingly this PR exposes some issues in handling empty column via ubsan.