@trivialfis 2020-07-15
Maybe dask is leaking file descriptor.