@trivialfis 2020-08-10
Dropped some optimizations to simplify the code, mostly affecting sparse datasets like URL.  Will do a new bench but I think the impact should be small.