@trivialfis 2020-08-11
## Before
- URL
Build: 3.18824s, 1 calls @ 3188235us
Build: 3.24613s, 1 calls @ 3246126us
Build: 3.29904s, 1 calls @ 3299044us

Mean: 3.244469

- HIGGS
Build: 2.53432s, 1 calls @ 2534319us
Build: 2.5671s, 1 calls @ 2567101us
Build: 2.4364s, 1 calls @ 2436397us

Mean: 2.512606

## After
- URL
SketchOnDMatrix: 3.47337s, 1 calls @ 3473373us
SketchOnDMatrix: 3.36139s, 1 calls @ 3661388us
SketchOnDMatrix: 3.56527s, 1 calls @ 3565270us

Mean: 3.466676

- HIGGS
SketchOnDMatrix: 2.48182s, 1 calls @ 2481822us
SketchOnDMatrix: 2.40048s, 1 calls @ 2400475us
SketchOnDMatrix: 2.43417s, 1 calls @ 2434170us

Mean: 2.438823

So slightly faster on dense dataset and slightly slower on extreme sparse dataset.  I variance should be negligible compared to total training time, with the benefits that now all tests run through the pipeline and easier to add categorical data support.