@ShvetsKS 2020-08-11
@trivialfis sorry for long response.
Could you also run distributed version with dmlc-submit on single node.

Initially before unification we observed slowdown of 'GmatInitialization' step (https://github.com/dmlc/xgboost/blob/d268a2a4631077c21562b606680c94a743ab7b18/src/tree/updater_quantile_hist.cc#L86) 
with configurations 1x56 (single worker with 56 available  threads) vs 7x8 (7 workers with 8 available threads): 19.37s vs 5.7s. 

It would be good to be sure that no regression was introduced for distributed case also.