@trivialfis 2020-08-11
Let me find a machine with that many threads.

> 19.37s vs 5.7s.

You mean after this PR the code slowed down for 4 times?