@trivialfis 2020-08-11
@ShvetsKS Speaking of distributed training, as we are merging rabit into XGBoost https://github.com/dmlc/xgboost/issues/5995 .  I think you can choose other backends in the future like UCX for better CPU perf if model recovery is not of concern.  Feel free to join the discussion and development.