@ShvetsKS 2020-08-11
Seems it could be even reproduced on similar configuration: 1x24 vs 4x6 for example.
Described slowdown was observed without this PR. I think performance check for distributed case is also needed to be sure that no other additional regression was introduced.