@trivialfis 2020-08-11
@ShvetsKS I did a quick run with dmlc-submit on HIGGS with a 4x4 configuration.

```
* With submit
** Before
GmatInitialization: 3.96191s, 1 calls @ 3961914us
GmatInitialization: 4.06724s, 1 calls @ 4067237us
GmatInitialization: 4.04202s, 1 calls @ 4042019us

** After
GmatInitialization: 4.0002s, 1 calls @ 4200197us
GmatInitialization: 4.06724s, 1 calls @ 4067243us
GmatInitialization: 3.98779s, 1 calls @ 3987792us
```

If there's slow down on GMatInit, I don't think it comes from sketching as I'm the last guy who made changes to it, and it has been a while now.