@trivialfis 2020-08-11
Result with setting `nthread` to 4 explicitly in above quick run script, value come from rank 0:

```
** Before

GmatInitialization: 3.83843s, 1 calls @ 3838427us
GmatInitialization: 3.89644s, 1 calls @ 3896437us
GmatInitialization: 3.90035s, 1 calls @ 3900354us

** After
```