@trivialfis 2020-08-11
@ShvetsKS  I summarized the perf I got so far.  Could you please take a look?