@ShvetsKS 2020-08-11
@trivialfis thanks a lot for information providing.
When we considered 1x56 or 1x16 configuration I mean this configuration:
./dmlc-core/tracker/dmlc-submit --cluster=local --num-workers=1 --worker-cores=16 python higgs-perf.py

I don't have any other concerns, look good generally.