@trivialfis 2020-08-11
@ShvetsKS I think the perf in description page should show the perf of single node vs multi node.