@trivialfis 2020-08-11
@ShvetsKS  Could you please explicitly approve the PR?  In general I think it would be best for me to have confirmation from you guys when I make big change to CPU code path.